DCS-1279: changes following PR comments
TransferResource deleted and content carried over to PrisonResource
Endpoint changed
